wm에 대한 이해 후 코드 수정(tooltip)
